Add/cart activity

+++ android/app/src/main/java/android/jodern/app/util/Utils.java
@@ -3,7 +3,7 @@ package android.jodern.app.util;
 import java.text.DecimalFormat;
 
 public class Utils {
-    public static String vndFormatPrice(int price) {
+    public static String vndFormatPrice(Long price) {
         DecimalFormat formatter = new DecimalFormat("###,###,###");
         return formatter.format(price) + " VNĐ";
     }
